+++ onemapsg/utils.py
@@ -12,7 +12,9 @@ from urllib.parse import urlencode
 import requests
 
 from onemapsg.api import API
-from onemapsg.response import Response, RouteResult, SearchResult
+from onemapsg.response import (
+    GeocodeInfo, Response, RouteResult, SearchResult
+)
 
 SAFE_METHODS = ['get', 'options']
 
@@ -69,6 +71,38 @@ def construct_search_query(search_val, return_geometry,
 def get_search_class():
     """Returns SearchResult class."""
     return SearchResult
+
+
+def validate_address_type(address_type: str) -> str:
+    if address_type.lower() not in ['all', 'hdb']:
+        raise ValueError(
+            'Invalid `addressType` value - can only be `HDB` or `All`'
+        )
+    return address_type.lower()
+
+
+def construct_reverse_geocode_svy21_query(location: (float, float),
+                                          token: str,
+                                          buffer: int=10,
+                                          address_type: str='all',
+                                          other_features: bool=False) -> str:
+    """Constructs Reverse Geocode (SVY21) query URL compliant with
+    OneMap's requirements."""
+    search_params = {
+        'location': ','.join([str(loc).strip() for loc in location]),
+        'token': token,
+        'buffer': buffer,
+        'addressType': validate_address_type(address_type),
+        'otherFeatures': 'Y' if other_features else 'N'
+    }
+    search_params = urlencode(search_params, safe=',:-')
+    search_url = f'{API.reverse_geocode_svy21}?{search_params}'
+    return search_url
+
+
+def get_reverse_geocode_svy21_class() -> GeocodeInfo:
+    """Returns GeocodeInfo class."""
+    return GeocodeInfo
 
 
 def construct_route_query(start, end, route_type,