+++ tests/test_utils.py
@@ -5,13 +5,34 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from onemapsg import status
-from onemapsg.response import Response
+from onemapsg.response import GeocodeInfo, Response, RouteResult, SearchResult
 from onemapsg.utils import (
-    construct_route_query, construct_search_query, make_request, to_dict
+    coerce_response,
+    construct_reverse_geocode_svy21_query,
+    construct_route_query, construct_search_query,
+    get_route_class,
+    get_search_class,
+    get_reverse_geocode_svy21_class,
+    make_request, to_dict,
+    validate_address_type
 )
 
 
+def test_coerce_response():
+    """Should correctly create a class object from dictionary."""
+    class TestClass:
+        def __init__(self, a, b, c):
+            self.a = a
+            self.b = b
+            self.c = c
+    instance = coerce_response(TestClass, {'a': 1, 'b': 2, 'c': 3})
+    assert instance.a == 1
+    assert instance.b == 2
+    assert instance.c == 3
+
+
 def test_to_dict():
+    """Should return correctly constructed dictionary from class."""
     class TestClass:
         def __init__(self, a, b, c):
             self.a = a
@@ -81,6 +102,12 @@ def test_construct_search_query():
                    'pageNum=1')
 
 
+def test_get_search_class():
+    """Should return SearchResult class."""
+    klass = get_search_class()
+    assert klass == SearchResult
+
+
 def test_construct_route_query():
     """Should construct correct route query."""
     url = construct_route_query('1.12,3.21', '1.02,1.05', 'drive',
@@ -107,3 +134,44 @@ def test_construct_route_query():
                    'time=15:30:00&'
                    'mode=BUS&'
                    'token=sometoken')
+
+
+def test_get_route_class():
+    """Should return RouteResult class."""
+    klass = get_route_class()
+    assert klass == RouteResult
+
+
+def test_validate_address_type():
+    """Should correctly validate address type for reverse geocode."""
+    with pytest.raises(ValueError) as err:
+        validate_address_type('badtype')
+        assert str(err) == 'Invalid `addressType` value - can only be `HDB` or `All`'
+    validated = validate_address_type('all')
+    assert validated == 'all'
+    validated = validate_address_type('hdb')
+    assert validated == 'hdb'
+    validated = validate_address_type('ALL')
+    assert validated == 'all'
+    validated = validate_address_type('HDB')
+    assert validated == 'hdb'
+
+
+def test_construct_reverse_geocode_svy21_query():
+    """Should construct correct reverse geocode query."""
+    location = (24291.97788882387, 31373.0117224489)
+    token = 'sometoken'
+    url = construct_reverse_geocode_svy21_query(location, token, other_features=True)
+    assert url == ('https://developers.onemap.sg/'
+                   'privateapi/commonsvc/revgeocodexy?'
+                   'location=24291.97788882387,31373.0117224489&'
+                   'token=sometoken&'
+                   'buffer=10&'
+                   'addressType=all&'
+                   'otherFeatures=Y')
+
+
+def test_get_reverse_geocode_svy21_class():
+    """Should return GeocodeInfo class."""
+    klass = get_reverse_geocode_svy21_class()
+    assert klass == GeocodeInfo